Finished one time payment for now. Split inputs and calculations from each other.
